chore: standardize repository constructor parameters and localization
