+++ .fvmrc
@@ -1,3 +1,3 @@
 {
-  "flutter": "3.44.2"
+  "flutter": "stable"
 }
