+++ .vscode/settings.json
@@ -50,5 +50,5 @@
     "**/.git": true,
     "flutter/**": true
   },
-  "dart.flutterSdkPath": ".fvm/versions/3.44.2"
+  "dart.flutterSdkPath": ".fvm/versions/stable"
 }
