--- lib/src/features/account/domain/repositories/account_repository.dart
@@ -1,33 +1,0 @@
-import 'package:thunder/src/foundation/primitives/primitives.dart';
-import 'package:thunder/src/features/account/domain/models/account_settings_update.dart';
-import 'package:thunder/src/features/account/domain/models/account_media.dart';
-
-//// Interface for an account repository
-abstract class AccountRepository {
-  /// Login to the account's home instance.
-  Future<String?> login({required String username, required String password, String? totp});
-
-  /// Logout the authenticated account on its home instance.
-  Future<void> logout();
-
-  /// Fetches the user's subscribed communities.
-  Future<List<ThunderCommunity>> subscriptions();
-
-  /// Fetches the user's media.
-  Future<ThunderPage<AccountMediaItem>> media({int? page, int? limit});
-
-  /// Saves the user's settings.
-  Future<void> saveSettings(AccountSettingsUpdate update);
-
-  /// Imports the settings to the user's profile.
-  Future<bool> importSettings(String settings);
-
-  /// Exports the user's settings.
-  Future<dynamic> exportSettings();
-
-  /// Upload an image.
-  Future<String> uploadImage(String filePath);
-
-  /// Delete an uploaded image.
-  Future<void> deleteImage({required String file, String? token});
-}
