+++ lib/src/features/account/presentation/state/profile_bloc.dart
@@ -4,6 +4,7 @@ import 'package:bloc/bloc.dart';
 import 'package:bloc_concurrency/bloc_concurrency.dart';
 import 'package:equatable/equatable.dart';
 import 'package:stream_transform/stream_transform.dart';
+import 'package:thunder/src/foundation/services/localization_service.dart';
 
 import 'package:thunder/src/foundation/utils/utils.dart';
 import 'package:thunder/src/foundation/primitives/primitives.dart';
