+++ lib/src/features/comment/comment.dart
@@ -8,5 +8,4 @@ export 'data/models/comment_node.dart';
 export 'domain/enums/comment_action.dart';
 export 'package:thunder/src/foundation/primitives/models/thunder_comment.dart';
 export 'domain/models/comment_page.dart';
-export 'domain/repositories/comment_repository.dart';
-export 'data/repositories/comment_repository_impl.dart';
+export 'data/repositories/comment_repository.dart';
