+++ lib/src/features/comment/data/repositories/comment_repository.dart
@@ -1,14 +1,73 @@
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
 
-import 'package:thunder/src/foundation/config/global_context.dart';
-import 'package:thunder/src/foundation/primitives/primitives.dart';
-import 'package:thunder/src/foundation/networking/networking.dart';
-import 'package:thunder/src/features/account/account.dart';
-import 'package:thunder/src/features/comment/comment.dart';
+import 'package:thunder/src/foundation/foundation.dart';
+import 'package:thunder/src/features/comment/domain/models/comment_page.dart';
 
-/// Implementation of [CommentRepository]
+/// Interface for a comment repository
+abstract class CommentRepository {
+  /// Fetches a comment by its ID
+  Future<ThunderComment> getComment(int commentId);
+
+  /// Fetches comments for a post
+  Future<CommentPage> getComments({
+    required int postId,
+    int? parentId,
+    int? page,
+    String? cursor,
+    CommentSortType? commentSortType,
+    int? maxDepth,
+    int? limit,
+    int? communityId,
+  });
+
+  /// Creates a new comment
+  Future<ThunderComment> create({
+    required int postId,
+    required String content,
+    int? parentId,
+    int? languageId,
+  });
+
+  /// Edits an existing comment
+  Future<ThunderComment> edit({
+    required int commentId,
+    required String content,
+    int? languageId,
+  });
+
+  /// Votes on a comment
+  Future<ThunderComment> vote(ThunderComment comment, int score);
+
+  /// Saves or unsaves a comment
+  Future<ThunderComment> save(ThunderComment comment, bool save);
+
+  /// Deletes a comment
+  Future<ThunderComment> delete(ThunderComment comment, bool deleted);
+
+  /// Reports a comment
+  Future<void> report(int commentId, String reason);
+
+  /// Creates a placeholder comment from the given parameters. This is mainly used to display a preview of the comment
+  /// with the applied settings on Settings -> Appearance -> Comments page.
+  Future<ThunderComment> createExample({
+    int? id,
+    String? path,
+    String? commentContent,
+    int? commentCreatorId,
+    int? commentScore,
+    int? commentUpvotes,
+    int? commentDownvotes,
+    DateTime? commentPublished,
+    int? commentChildCount,
+    String? personName,
+    String? personAvatar,
+    bool? isPersonAdmin,
+    bool? isBotAccount,
+    bool? saved,
+  });
+}
+
+/// Implementation of [CommentRepository] using the unified API client
 class CommentRepositoryImpl implements CommentRepository {
   /// The account to use for methods invoked in this repository
   final Account account;
@@ -16,10 +75,18 @@ class CommentRepositoryImpl implements CommentRepository {
   /// The API client to use for the repository
   final ThunderApiClient _api;
 
+  /// The localization service to use for user-facing errors
+  final LocalizationService _localization;
+
   /// Creates a new CommentRepositoryImpl.
   ///
-  /// An optional [api] client can be provided for testing.
-  CommentRepositoryImpl({required this.account, ThunderApiClient? api}) : _api = api ?? ApiClientFactory.create(account, debug: kDebugMode);
+  /// An optional [api] client and [localization] can be provided for testing.
+  CommentRepositoryImpl({
+    required this.account,
+    ThunderApiClient? api,
+    LocalizationService localization = const ThunderLocalizationService(),
+  })  : _api = api ?? ApiClientFactory.create(account, debug: kDebugMode),
+        _localization = localization;
 
   @override
   Future<ThunderComment> getComment(int commentId) async {
@@ -61,7 +128,7 @@ class CommentRepositoryImpl implements CommentRepository {
     int? parentId,
     int? languageId,
   }) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.createComment(
@@ -78,7 +145,7 @@ class CommentRepositoryImpl implements CommentRepository {
     required String content,
     int? languageId,
   }) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.editComment(
@@ -90,7 +157,7 @@ class CommentRepositoryImpl implements CommentRepository {
 
   @override
   Future<ThunderComment> vote(ThunderComment comment, int score) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.voteComment(commentId: comment.id, score: score);
@@ -98,7 +165,7 @@ class CommentRepositoryImpl implements CommentRepository {
 
   @override
   Future<ThunderComment> save(ThunderComment comment, bool save) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.saveComment(commentId: comment.id, save: save);
@@ -106,7 +173,7 @@ class CommentRepositoryImpl implements CommentRepository {
 
   @override
   Future<ThunderComment> delete(ThunderComment comment, bool deleted) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.deleteComment(commentId: comment.id, deleted: deleted);
@@ -114,7 +181,7 @@ class CommentRepositoryImpl implements CommentRepository {
 
   @override
   Future<void> report(int commentId, String reason) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     await _api.reportComment(commentId: commentId, reason: reason);
