+++ lib/src/features/comment/presentation/state/create_comment_cubit.dart
@@ -3,11 +3,11 @@ import 'package:flutter/foundation.dart';
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 
-import 'package:thunder/src/foundation/contracts/contracts.dart';
 import 'package:thunder/src/foundation/errors/errors.dart';
 import 'package:thunder/src/features/account/account.dart';
 import 'package:thunder/src/features/comment/comment.dart';
 import 'package:thunder/src/foundation/networking/networking.dart';
+import 'package:thunder/src/foundation/services/localization_service.dart';
 
 part 'create_comment_state.dart';
