+++ lib/src/features/comment/presentation/utils/comment_utils.dart
@@ -6,7 +6,7 @@ import 'package:thunder/src/features/account/account.dart';
 import 'package:thunder/src/features/comment/data/models/comment_node.dart';
 import 'package:thunder/src/foundation/primitives/models/thunder_comment.dart';
 import 'package:thunder/src/foundation/primitives/models/vote_state.dart';
-import 'package:thunder/src/features/comment/data/repositories/comment_repository_impl.dart';
+import 'package:thunder/src/features/comment/data/repositories/comment_repository.dart';
 import 'package:thunder/src/features/comment/domain/enums/comment_action.dart';
 
 // Optimistically updates a comment
