+++ lib/src/features/notification/data/repositories/notification_repository.dart
@@ -1,10 +1,6 @@
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
 
-import 'package:thunder/src/foundation/contracts/contracts.dart';
-import 'package:thunder/src/foundation/primitives/primitives.dart';
-import 'package:thunder/src/foundation/networking/networking.dart';
+import 'package:thunder/src/foundation/foundation.dart';
 import 'package:thunder/src/features/notification/domain/models/unread_notifications_count.dart';
 
 /// Interface for a notification repository
@@ -57,7 +53,7 @@ abstract class NotificationRepository {
   Future<void> markAllNotificationsAsRead();
 }
 
-/// Implementation of [NotificationRepository]
+/// Implementation of [NotificationRepository] using the unified API client
 class NotificationRepositoryImpl implements NotificationRepository {
   /// The account to use for methods invoked in this repository
   final Account account;
@@ -66,17 +62,17 @@ class NotificationRepositoryImpl implements NotificationRepository {
   final ThunderApiClient _api;
 
   /// The localization service to use for user-facing errors
-  final LocalizationService _localizationService;
+  final LocalizationService _localization;
 
   /// Creates a new NotificationRepositoryImpl.
   ///
-  /// An optional [api] client can be provided for testing.
+  /// An optional [api] client and [localization] can be provided for testing.
   NotificationRepositoryImpl({
     required this.account,
     ThunderApiClient? api,
-    LocalizationService localizationService = const GlobalContextLocalizationService(),
+    LocalizationService localization = const ThunderLocalizationService(),
   })  : _api = api ?? ApiClientFactory.create(account, debug: kDebugMode),
-        _localizationService = localizationService;
+        _localization = localization;
 
   @override
   Future<List<ThunderComment>> replies({
@@ -85,15 +81,15 @@ class NotificationRepositoryImpl implements NotificationRepository {
     CommentSortType sort = CommentSortType.new_,
     int page = 1,
   }) async {
-    if (account.anonymous) {
-      throw Exception(_localizationService.l10n.userNotLoggedIn);
-    }
+    final l10n = _localization.l10n;
+    if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
+
     return await _api.getCommentReplies(page: page, limit: limit, sort: sort, unread: unread);
   }
 
   @override
   Future<void> markReplyAsRead({required int replyId, bool read = true}) async {
-    final l10n = _localizationService.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     await _api.markCommentReplyAsRead(replyId: replyId, read: read);
@@ -106,7 +102,7 @@ class NotificationRepositoryImpl implements NotificationRepository {
     CommentSortType sort = CommentSortType.new_,
     int page = 1,
   }) async {
-    final l10n = _localizationService.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.getCommentMentions(page: page, limit: limit, sort: sort, unread: unread);
@@ -114,7 +110,7 @@ class NotificationRepositoryImpl implements NotificationRepository {
 
   @override
   Future<void> markMentionAsRead({required int mentionId, bool read = true}) async {
-    final l10n = _localizationService.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     await _api.markCommentMentionAsRead(mentionId: mentionId, read: read);
@@ -126,7 +122,7 @@ class NotificationRepositoryImpl implements NotificationRepository {
     int limit = 50,
     int page = 1,
   }) async {
-    final l10n = _localizationService.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.getPrivateMessages(page: page, limit: limit, unread: unread);
@@ -134,7 +130,7 @@ class NotificationRepositoryImpl implements NotificationRepository {
 
   @override
   Future<void> markMessageAsRead({required int notificationId, bool read = true}) async {
-    final l10n = _localizationService.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     await _api.markPrivateMessageAsRead(notificationId: notificationId, read: read);
@@ -142,7 +138,7 @@ class NotificationRepositoryImpl implements NotificationRepository {
 
   @override
   Future<UnreadNotificationsCount> unreadNotificationsCount() async {
-    final l10n = _localizationService.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     final response = await _api.unreadCount();
@@ -155,7 +151,7 @@ class NotificationRepositoryImpl implements NotificationRepository {
 
   @override
   Future<void> markAllNotificationsAsRead() async {
-    final l10n = _localizationService.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     await _api.markAllNotificationsAsRead();
