+++ lib/src/features/post/data/repositories/post_repository.dart
@@ -1,12 +1,6 @@
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
 
-import 'package:thunder/src/foundation/config/global_context.dart';
-import 'package:thunder/src/foundation/errors/errors.dart';
-import 'package:thunder/src/foundation/networking/networking.dart';
-import 'package:thunder/src/foundation/primitives/primitives.dart';
-import 'package:thunder/src/features/account/account.dart';
+import 'package:thunder/src/foundation/foundation.dart';
 import 'package:thunder/src/features/post/post.dart';
 
 /// Interface for a post repository
@@ -100,7 +94,7 @@ abstract class PostRepository {
   Future<void> report(int postId, String reason);
 }
 
-/// Implementation of [PostRepository]
+/// Implementation of [PostRepository] using the unified API client
 class PostRepositoryImpl implements PostRepository {
   /// The account to use for methods invoked in this repository
   final Account account;
@@ -108,10 +102,18 @@ class PostRepositoryImpl implements PostRepository {
   /// The API client to use for the repository
   final ThunderApiClient _api;
 
+  /// The localization service to use for user-facing errors
+  final LocalizationService _localization;
+
   /// Creates a new PostRepositoryImpl.
   ///
-  /// An optional [api] client can be provided for testing.
-  PostRepositoryImpl({required this.account, ThunderApiClient? api}) : _api = api ?? ApiClientFactory.create(account, debug: kDebugMode);
+  /// An optional [api] client and [localization] can be provided for testing.
+  PostRepositoryImpl({
+    required this.account,
+    ThunderApiClient? api,
+    LocalizationService localization = const ThunderLocalizationService(),
+  })  : _api = api ?? ApiClientFactory.create(account, debug: kDebugMode),
+        _localization = localization;
 
   @override
   Future<Map<String, dynamic>?> getPost(int postId, {int? commentId}) async {
@@ -183,7 +185,7 @@ class PostRepositoryImpl implements PostRepository {
     int? postIdBeingEdited,
     int? languageId,
   }) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     ThunderPost response;
@@ -222,7 +224,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<ThunderPost> vote(ThunderPost post, int score) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     final response = await _api.votePost(postId: post.id, score: score);
@@ -231,7 +233,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<ThunderPost> save(ThunderPost post, bool save) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     final response = await _api.savePost(postId: post.id, save: save);
@@ -240,7 +242,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<bool> read(int postId, bool read) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.readPost(postIds: [postId], read: read);
@@ -248,7 +250,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<List<int>> readMultiple(List<int> postIds, bool read) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     final success = await _api.readPost(postIds: postIds, read: read);
@@ -257,7 +259,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<bool> hide(int postId, bool hide) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     if (!_api.supportsHidePosts) {
@@ -269,7 +271,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<bool> delete(int postId, bool delete) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.deletePost(postId: postId, deleted: delete);
@@ -277,7 +279,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<bool> lock(int postId, bool lock) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.lockPost(postId: postId, locked: lock);
@@ -285,7 +287,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<bool> pinCommunity(int postId, bool pin) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.pinPost(postId: postId, pinned: pin);
@@ -293,7 +295,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<bool> remove(int postId, bool remove, String reason) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     return await _api.removePost(postId: postId, removed: remove, reason: reason);
@@ -301,7 +303,7 @@ class PostRepositoryImpl implements PostRepository {
 
   @override
   Future<void> report(int postId, String reason) async {
-    final l10n = GlobalContext.l10n;
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
     await _api.reportPost(postId: postId, reason: reason);
