+++ lib/src/features/session/data/repositories/session_repository.dart
@@ -5,7 +5,34 @@ import 'package:thunder/src/foundation/config/config.dart';
 import 'package:thunder/src/foundation/contracts/account.dart';
 import 'package:thunder/src/foundation/persistence/persistence.dart';
 import 'package:thunder/src/foundation/primitives/primitives.dart';
-import 'package:thunder/src/features/session/domain/repositories/session_repository.dart';
+
+abstract class SessionRepository {
+  Future<Account> bootstrap();
+
+  Future<List<Account>> getAuthenticatedSessions();
+
+  Future<List<Account>> getAnonymousSessions();
+
+  Future<List<Account>> getSessions();
+
+  Future<Account?> getSessionByKey(String sessionKey);
+
+  Future<void> setActiveSession(Account account);
+
+  Future<void> clearActiveSession();
+
+  Future<Account?> addAuthenticatedSession(Account account);
+
+  Future<Account?> addAnonymousSession(Account account);
+
+  Future<void> removeSession(String sessionKey);
+
+  /// Atomically updates the display order of authenticated and anonymous profiles.
+  Future<void> updateSessionOrder({
+    required List<Account> authenticatedSessions,
+    required List<Account> anonymousSessions,
+  });
+}
 
 class PersistentSessionRepository implements SessionRepository {
   const PersistentSessionRepository();
