+++ lib/src/features/session/presentation/state/session_bloc.dart
@@ -1,12 +1,13 @@
-import 'package:bloc/bloc.dart';
-import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 
-import 'package:thunder/src/foundation/contracts/contracts.dart';
+import 'package:bloc/bloc.dart';
+import 'package:equatable/equatable.dart';
+
 import 'package:thunder/src/foundation/utils/cache/platform_version_cache.dart';
 import 'package:thunder/src/features/account/account.dart';
 import 'package:thunder/src/features/instance/instance.dart';
-import 'package:thunder/src/features/session/domain/repositories/session_repository.dart';
+import 'package:thunder/src/features/session/data/repositories/session_repository.dart';
+import 'package:thunder/src/foundation/services/localization_service.dart';
 
 part 'session_event.dart';
 part 'session_state.dart';
