+++ lib/src/features/session/session.dart
@@ -1,9 +1,8 @@
-export 'domain/repositories/session_repository.dart';
+export 'data/repositories/session_repository.dart';
 export 'domain/models/feature_account.dart';
 export 'domain/models/feature_account_resolution_request.dart';
 export 'domain/models/feature_account_resolved_content.dart';
 export 'domain/services/feature_account_content_resolver.dart';
-export 'data/repositories/persistent_session_repository.dart';
 export 'presentation/utils/effective_account_context.dart';
 export 'presentation/state/feature_account_cubit.dart';
 export 'presentation/state/session_bloc.dart';
