+++ lib/src/features/user/data/repositories/user_repository.dart
@@ -1,15 +1,9 @@
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
 
-import 'package:thunder/src/foundation/config/global_context.dart';
-import 'package:thunder/src/foundation/primitives/primitives.dart';
-import 'package:thunder/src/foundation/networking/networking.dart';
-import 'package:thunder/src/features/account/account.dart';
+import 'package:thunder/src/foundation/foundation.dart';
 
-/// Interface for a user repository
 abstract class UserRepository {
-  /// Fetches a user by its ID
+  /// Fetches a user by their id or username
   Future<Map<String, dynamic>?> getUser({
     int? userId,
     String? username,
@@ -21,8 +15,8 @@ abstract class UserRepository {
     bool? includeContent,
   });
 
-  /// Blocks or unblocks a person
-  Future<ThunderUser> block(int personId, bool block);
+  /// Blocks or unblocks a user
+  Future<ThunderUser> blockUser(int userId, bool block);
 }
 
 /// Implementation of [UserRepository] using the unified API client
@@ -33,10 +27,18 @@ class UserRepositoryImpl implements UserRepository {
   /// The API client to use for the repository
   final ThunderApiClient _api;
 
+  /// The localization service to use for user-facing errors
+  final LocalizationService _localization;
+
   /// Creates a new UserRepositoryImpl.
   ///
-  /// An optional [api] client can be provided for testing.
-  UserRepositoryImpl({required this.account, ThunderApiClient? api}) : _api = api ?? ApiClientFactory.create(account, debug: kDebugMode);
+  /// An optional [api] client and [localization] can be provided for testing.
+  UserRepositoryImpl({
+    required this.account,
+    ThunderApiClient? api,
+    LocalizationService localization = const ThunderLocalizationService(),
+  })  : _api = api ?? ApiClientFactory.create(account, debug: kDebugMode),
+        _localization = localization;
 
   @override
   Future<Map<String, dynamic>?> getUser({
@@ -71,10 +73,10 @@ class UserRepositoryImpl implements UserRepository {
   }
 
   @override
-  Future<ThunderUser> block(int personId, bool block) async {
-    final l10n = GlobalContext.l10n;
+  Future<ThunderUser> blockUser(int userId, bool block) async {
+    final l10n = _localization.l10n;
     if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
 
-    return await _api.blockUser(userId: personId, block: block);
+    return await _api.blockUser(userId: userId, block: block);
   }
 }
