+++ lib/src/foundation/contracts/contracts.dart
@@ -1,7 +1,6 @@
 export 'account.dart';
 export 'connectivity_service.dart';
 export 'deep_link_service.dart';
-export 'localization_service.dart';
 export 'notification_service.dart';
 export 'platform_detection_service.dart';
 export 'preferences_store.dart';
