+++ lib/src/foundation/services/services.dart
@@ -1,0 +1,1 @@
+export 'localization_service.dart';
