+++ test/thunder_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  testWidgets('Sample Test', (WidgetTester tester) async {});
+  testWidgets('Test', (WidgetTester tester) async {});
 }
